Defer restoration data saves
Fix the tests
